backend/requirements.txt
fastapi
uvicorn
motor>=3.0.0
pydantic-settings
bcrypt
pyjwt
python-multipart
requests
dnspython
pydantic[email]
pymongo>=4.0.0
certifi
# ML dependencies for stutter detection and transcription
torch
transformers
librosa
soundfile
openai-whisper
